+++ apps/rig_api/lib/rig_api/controllers/fallback_controller.ex
@@ -1,0 +1,14 @@
+defmodule RigApi.FallbackController do
+  @moduledoc """
+  Translates controller action results into valid `Plug.Conn` responses.
+
+  See `Phoenix.Controller.action_fallback/1` for more details.
+  """
+  use RigApi, :controller
+
+  def call(conn, {:error, :not_found}) do
+    conn
+    |> put_status(:not_found)
+    |> render(RigApi.ErrorView, :"404")
+  end
+end