+++ apps/rig_outbound_gateway/lib/rig_outbound_gateway/application.ex
@@ -1,0 +1,16 @@
+defmodule RigOutboundGateway.Application do
+  @moduledoc false
+
+  use Application
+
+  alias RigOutboundGateway.Kafka
+
+  def start(_type, _args) do
+    children = [
+      Kafka.SupWrapper
+    ]
+
+    opts = [strategy: :one_for_one, name: RigOutboundGateway.Supervisor]
+    Supervisor.start_link(children, opts)
+  end
+end